backend/requirements.txt
ddgs==9.11.4
fastapi==0.135.2
httpx==0.28.1
langchain-core>=0.3.0
langchain-ollama>=0.2.0
numpy>=1.24.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==1.0.0
uvicorn==0.42.0
